Enable afterburner. Add total kills / deaths to final score screen. Remove PlayerData.

+++ src/Inferno/Game.Reactor.cpp
@@ -332,7 +332,7 @@ namespace Inferno::Game {
                 Game::FireWeaponInfo info = { .id = WeaponID::ReactorBlob, .gun = (uint8)gun, .customDir = &dir };
                 FireWeapon(reactor, info);
 
-                Game::PlayWeaponSound(WeaponID::ReactorBlob, 1, reactor, gun);
+                Game::PlayWeaponSound(WeaponID::ReactorBlob, 1, reactor, (uint8)gun);
 
                 // Randomly fire more blobs based on level number and difficulty
                 auto chance = 1.0f / ((float)Game::LevelNumber / 4 + 2);
